requirements.txt
flake8
Scrapy
selenium
python-dotenv
